feat: align new plant page with style guide
